+++ Commands/ClientSidePages/ConvertToClientSidePage.cs
@@ -69,6 +69,10 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
     Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""My post title"" -BlogPage -LogType Console -Overwrite -TargetWebUrl https://contoso.sharepoint.com/sites/targetmodernsite",
     Remarks = "Converts a blog page with a title starting with 'my post title' to a client side page in the https://contoso.sharepoint.com/sites/targetmodernsite site",
     SortOrder = 12)]
+    [CmdletExample(
+    Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""somepage.aspx"" -PublishingPage -Overwrite -TargetConnection $target -UserMappingFile c:\\temp\user_mapping_file.csv",
+    Remarks = "Converts a publishing page named 'somepage' to a client side page in the site specified by the TargetConnection connection. This allows to read a page in on-premises environment and create in another online locations including using specific user mappings between the two environments.",
+    SortOrder = 13)]
     public class ConvertToClientSidePage : PnPWebCmdlet
     {
         private static string rootFolder = "<root>";
@@ -172,8 +176,20 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
         [Parameter(Mandatory = false, HelpMessage = "Name for the target page (only applies when doing cross site page transformation)")]
         public string TargetPageName = "";
 
+        [Parameter(Mandatory = false, HelpMessage = "Folder to create the target page in (will be used in conjunction with auto-generated folders that ensure page name uniqueness)")]
+        public string TargetPageFolder = "";
+
         [Parameter(Mandatory = false, HelpMessage = "Optional connection to be used by the cmdlet. Retrieve the value for this parameter by either specifying -ReturnConnection on Connect-PnPOnline or by executing Get-PnPConnection.")] // do not remove '#!#99'
         public SPOnlineConnection TargetConnection = null;
+
+        [Parameter(Mandatory = false, HelpMessage = "Disables user mapping during transformation")]
+        public SwitchParameter SkipUserMapping = false;
+
+        [Parameter(Mandatory = false, HelpMessage = "Specifies a user mapping file")]
+        public string UserMappingFile = "";
+
+        [Parameter(Mandatory = false, HelpMessage = "Specifies a LDAP connection string e.g. LDAP://OU=Users,DC=Contoso,DC=local")]
+        public string LDAPConnectionString = "";
 
         protected override void ExecuteCmdlet()
         {
@@ -381,6 +397,10 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
                     SkipDefaultUrlRewrite = this.SkipDefaultUrlRewriting,
                     UrlMappingFile = this.UrlMappingFile,
                     AddTableListImageAsImageWebPart = this.AddTableListImageAsImageWebPart,
+                    SkipUserMapping = this.SkipUserMapping,
+                    UserMappingFile = this.UserMappingFile,
+                    LDAPConnectionString = this.LDAPConnectionString,
+                    TargetPageFolder = this.TargetPageFolder,
                 };
 
                 // Set mapping properties
@@ -431,6 +451,10 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
                     SkipDefaultUrlRewrite = this.SkipDefaultUrlRewriting,
                     UrlMappingFile = this.UrlMappingFile,
                     AddTableListImageAsImageWebPart = this.AddTableListImageAsImageWebPart,
+                    SkipUserMapping = this.SkipUserMapping,
+                    UserMappingFile = this.UserMappingFile,
+                    LDAPConnectionString = this.LDAPConnectionString,
+                    TargetPageFolder = this.TargetPageFolder,
                     ModernizationCenterInformation = new ModernizationCenterInformation()
                     {
                         AddPageAcceptBanner = this.AddPageAcceptBanner